App now keeps working even if API is down

+++ F1Driver/DataLayer/Classes/PopulateDatabase.cs
@@ -22,22 +22,38 @@ namespace DataLayer.Classes
         }
         public async Task<bool> Populate()
         {//Method that calls all the classes to fill the database
-            await ClearYearDrivers();
+                await ClearYear();
             PopulateDriversAndTeams getDriverData = new(_context);
             PopulateCircuits getCircuitData = new(_context);
             PopulateRaces getRaceData = new(_context);
-             await getCircuitData.InsertCircuits(await getCircuitData.GetCircuits());
-            await getDriverData.InsertConstrutors(await getDriverData.GetConstructors());
-            await getDriverData.LinkDriverToConstructor(await getDriverData.CreateDriver());
-            getRaceData.RaceHelper();
-            return true;
+            try
+            {
+                await getCircuitData.InsertCircuits(await getCircuitData.GetCircuits());
+                await getDriverData.InsertConstrutors(await getDriverData.GetConstructors());
+                await getDriverData.LinkDriverToConstructor(await getDriverData.CreateDriver());
+                getRaceData.RaceHelper();
+                return true;
+            }
+            catch (Exception e)
+            {//If the API is down go here.
+                Console.WriteLine(e);
+                return false;
+            }
         }
-       public Task<bool> ClearYearDrivers()
+        public async Task<bool> ClearYear()
         {//drops all drivers so old ones dont stay in the system
-            _context.Driver.FromSqlRaw("TRUNCATE driver").ToListAsync();
-            _context.Constructor.FromSqlRaw("TRUNCATE constructor").ToListAsync();
-            _context.SaveChangesAsync();
-            return Task.FromResult(true);
+            int date = DateTime.Now.Year;
+            List<RaceModel> races = await _context.Race.Where(race => race.Season != date).ToListAsync();
+            if (races.Count!=0)
+            {
+                await _context.Driver.FromSqlRaw("TRUNCATE driver").ToListAsync();
+                await _context.Circuit.FromSqlRaw("TRUNCATE circuit").ToListAsync();
+                await _context.Race.FromSqlRaw("TRUNCATE race").ToListAsync();
+                await _context.RaceResult.FromSqlRaw("TRUNCATE raceresult").ToListAsync();
+                await _context.Constructor.FromSqlRaw("TRUNCATE constructor").ToListAsync();
+                await _context.SaveChangesAsync();
+            }
+            return true;
         }
     }
 }
